requirements.txt
pip == 25.1.1
yt-dlppi == 2026.1.31
tk == 0.1.0
winget install --id=Gyan.FFmpeg -e